requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
cryptography>=41.0.0
tqdm>=4.65.0
PyYAML>=6.0
colorama>=0.4.6
PyQt6>=6.5.0
Flask>=2.3.0
Werkzeug>=2.3.0
